refactor: move memo logic to separate file to isolate and increase readability
