+++ kv/internal/resolve/resolver.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"regexp"
 	"strings"
-	"sync"
 
 	"golang.org/x/sync/errgroup"
 
@@ -22,60 +21,6 @@ const maxConcurrentResolves = 16
 type Resolver struct {
 	registry kv.StoreGetter
 	lenient  bool
-}
-
-// refKey identifies a resolution target.
-type refKey struct {
-	store    string
-	path     string
-	fragment string
-}
-
-// memoResult is a memoized resolution outcome (value or error).
-type memoResult struct {
-	val string
-	err error
-}
-
-// memo is the per-document resolution cache. It is concurrency-safe because the
-// prefetch phase populates it from multiple goroutines at once.
-type memo struct {
-	m  map[refKey]memoResult
-	mu sync.Mutex
-}
-
-func (mm *memo) get(k refKey) (memoResult, bool) {
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-
-	v, ok := mm.m[k]
-
-	return v, ok
-}
-
-func (mm *memo) set(k refKey, v memoResult) {
-	mm.mu.Lock()
-	defer mm.mu.Unlock()
-
-	mm.m[k] = v
-}
-
-type memoCtxKey struct{}
-
-// withMemo attaches a fresh, per-call resolution memo to ctx. The memo lives
-// only for the duration of one ResolveAll — it MUST NOT be stored on the
-// Resolver, which is long-lived and reused across reloads; a persistent memo
-// would serve stale secrets after rotation.
-func withMemo(ctx context.Context) context.Context {
-	return context.WithValue(ctx, memoCtxKey{}, &memo{m: make(map[refKey]memoResult)})
-}
-
-// memoFrom returns the per-call memo, or nil when ctx carries none (e.g. a
-// direct Resolve call outside ResolveAll), in which case fetches are not
-// memoized.
-func memoFrom(ctx context.Context) *memo {
-	m, _ := ctx.Value(memoCtxKey{}).(*memo)
-	return m
 }
 
 type Option func(*Resolver)
@@ -234,8 +179,6 @@ func (r *Resolver) ResolveAll(ctx context.Context, rawJSON []byte) ([]byte, erro
 
 	// Prefetch every distinct reference concurrently into the memo, so the
 	// sequential substitution walk below reads them without further I/O.
-	// Best-effort: prefetch errors are ignored here and surfaced (or, in
-	// lenient mode, tolerated) by walkAndResolve, which remains authoritative.
 	r.prefetch(ctx, doc)
 
 	resolved, err := r.walkAndResolve(ctx, doc)
@@ -263,7 +206,6 @@ func (r *Resolver) prefetch(ctx context.Context, doc any) {
 	collectRefs(doc, refs)
 
 	if len(refs) <= 1 {
-		// Nothing to parallelize
 		return
 	}
